TST - Moved NonDictMapping test util class to pandas/util/testing.py TST - Added a test of constructing a series from NonDictMapping
